Move TagsGallery component to tags component directory
